Add a rudimentary acceptance test
We add an acceptance test for ensuring that unmarshallable exceptions
are properly reported. Testing that feature is impossible from within
Minitest as we're interested in making an assertion on the output
produced by it.

We solved the problem by running a Minitest test file and comparing
its output against an excpeted output.

Last but not least, we updated the Gemfile.